requirements.txt
lxml==3.6.0
python-dateutil==2.5.3
requests==2.9.1
